requirements.txt
requests>=2.13.0
boto3>=1.4.6
python-dotenv>=0.7.1
